Added a code block type of `dotascii`, which will run the code block through `graph-easy` which accepts graphviz dot graphs and creates ascii art from them.  The ascii art is rendered as a code block.
`confluence-pusher` is supposed to handle any graphviz output and embed the image
in the resulting document, however, this doesn't seem to work.  It may be due to differences
between Confluence server and cloud.  Anyway, at the moment, ascii graphviz is the best
that can be done.

I believe future work is to create a png from graphviz, attach the file, and insert a link
to the attachment.